refacto navbar

+++ projects/cashier/src/app/game-cards/game-cards-editor/game-cards-editor.component.ts
@@ -1,13 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { GameCardService } from '../../game-card.service';
 import { CommonModule } from '@angular/common';
-import { GameCard, MAX_STAMPS } from '../game-card.interface';
+import { GameCard } from '../game-card.interface';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { GetGameCardsOwnersComponent } from '../get-game-cards-owners/get-game-cards-owners.component';
 import { Member } from '../../../../../common/member.interface';
-import { MembersService } from '../../../../../admin-dashboard/src/app/members/service/members.service';
-import { switchMap, tap } from 'rxjs';
 import { EditGameCardComponent } from '../edit-game-card/edit-game-card.component';
+import { GetConfirmationComponent } from '../../modals/get-confirmation/get-confirmation.component';
 
 @Component({
   selector: 'app-game-cards-editor',
@@ -19,6 +18,7 @@ import { EditGameCardComponent } from '../edit-game-card/edit-game-card.componen
 export class GameCardsEditorComponent implements OnInit {
 
   cards: GameCard[] = [];
+  total_asset: number = 0;
 
   constructor(
     private gameCardService: GameCardService,
@@ -30,6 +30,7 @@ export class GameCardsEditorComponent implements OnInit {
 
     this.gameCardService.gameCards.subscribe(cards => {
       this.cards = cards;
+      this.total_asset = cards.reduce((acc, card) => acc + card.initial_qty-card.stamps.length, 0);
     });
   }
 
@@ -39,22 +40,31 @@ stamps_number(card: GameCard): number {
 
   createGameCard() {
     const modalRef = this.modalService.open(GetGameCardsOwnersComponent, { centered: true });
-    modalRef.result.then((response: Member[]) => {
+    modalRef.result.then((response: { owners:Member[],qty:number}) => {
       if (response) {
-        const owners: Member[] = response;
+        const owners: Member[] = response.owners;
+        const qty: number = response.qty;
 
-        this.gameCardService.createCard(owners);
+        this.gameCardService.createCard(owners,qty);
       }
     });
   }
 
-  // Example method to update an existing game card
   updateGameCard(card: GameCard) {
     this.gameCardService.updateCard(card)
   }
 
   deleteGameCard(card: GameCard) {
-    this.gameCardService.deleteCard(card)
+const modalRef = this.modalService.open(GetConfirmationComponent, { centered: true });
+const owners = card.owners.reduce((acc, owner) => acc + owner.firstname + ' ' + owner.lastname.toUpperCase() + ' ', '');
+    modalRef.componentInstance.title = `Suppression d'une carte d'admission `; 
+    modalRef.componentInstance.subtitle = `détenteur(s) :${owners} `; 
+    modalRef.result.then((answer: boolean) => {
+      if (answer) {
+        this.gameCardService.deleteCard(card)
+      }
+    });
+
   }
